Add vue-org-tree, vue-organization-chart, and related routes for organization chart feature
The diff includes the addition of the "vue-org-tree" and "vue-organization-chart" packages to the project's dependencies in package.json, as well as the creation of new Vue components and routes for displaying an organization chart. The chart component is added to the router configuration in index.js, and a new route for accessing the chart is created. Additionally, a new route for accessing a template with aliases is added. The chart and templateAlis components are located in the views directory.