"Update template initialization to add new templates at the beginning of the array"
This commit updates the `initAuth` method to add new templates at the beginning of the `templates` array instead of the end. This is done by replacing the `push` method with the `unshift` method when adding the new template to the array. This change affects both logged-in and logged-out users. Additionally, a newline and a comment have been added to the code for readability.